refactor: move api calls of Club page to clubAPI.ts
