+++ src/modules/Dashboard/modules/pages/Club/ClubBanner.tsx
@@ -3,6 +3,7 @@ import BannerImg from "../../../../../assets/Study abroad-pana.png"
 import './Banner.scss'
 import { privateGateway } from '../../../../../services/apiGateway'
 import { bannerRoutes } from '../../../../../services/urls'
+import { fetchInstitutionStatusCount } from './clubAPI'
 const ClubBanner: FC<{ updated: boolean }> = ({ updated }) => {
     const [count, setCount] = useState<CountResponse>(initialState)
     useEffect(() => {
@@ -36,7 +37,7 @@ const initialState: CountResponse = {
     "Execom Formed": 0,
     total: 0,
 };
-interface CountResponse {
+export interface CountResponse {
     Identified: number;
     Confirmed: number;
     "Connection established": number;
@@ -45,10 +46,5 @@ interface CountResponse {
     "Execom Formed": number;
     total: number;
 }
-const fetchInstitutionStatusCount = async (setCount: Dispatch<SetStateAction<CountResponse>>) => {
-    privateGateway.get(`${bannerRoutes.clubBanner}`)
-        .then(res => res.data.response)
-        .then(res => setCount(res))
-        .catch(err => console.log(err))
-}
+
 export default ClubBanner
