+++ src/modules/Dashboard/modules/pages/Club/ClubSetup.tsx
@@ -8,6 +8,7 @@ import '../../components/Setup.scss'
 import { initialState, selectProps } from '../../utils/setupUtils'
 import * as yup from 'yup'
 import { Error, Success, showAlert } from '../../../components/Error/Alerts'
+import { createClub, fetchDistricts, fetchcolleges } from './clubAPI'
 interface ClubSetupProps {
     setViewSetup: Dispatch<SetStateAction<boolean>>
     updateClubData: Function
@@ -102,38 +103,5 @@ const ClubSetup: FC<ClubSetupProps> = ({ setViewSetup, updateClubData }) => {
         </div>
     )
 }
-function fetchDistricts(setData: Dispatch<SetStateAction<selectProps[]>>) {
-    privateGateway.get(setupRoutes.district.list)
-        .then(res => res.data.response.districts)
-        .then(data => setData(data))
-        .catch(err => console.log(err))
-}
 
-function fetchcolleges(setData: Dispatch<SetStateAction<selectProps[]>>, districtName: string) {
-    const reqData: any = {
-        district: districtName
-    }
-    privateGateway.post(setupRoutes.district.college, reqData)
-        .then(res => res.data.response.institutions)
-        .then(data => {
-            setData(data.map((item: any) => ({ id: item.id, name: item.title, })))
-        })
-        .catch(err => console.log(err))
-}
-function createClub<postDataProps>(postData: postDataProps, update: Function, setViewSetup: Dispatch<SetStateAction<boolean>>, setSuccessMessage: Dispatch<SetStateAction<string>>,
-    setErrorMessage: Dispatch<SetStateAction<string>>) {
-    privateGateway.post(setupRoutes.club.create, postData)
-        .then(res => {
-            update()
-            showAlert(res?.data?.message?.general[0], setSuccessMessage)
-            console.log('Success :', res?.data?.message?.general[0])
-            setTimeout(() => {
-                setViewSetup(false)
-            }, 3000)
-        })
-        .catch(err => {
-            showAlert(err?.response?.data?.message?.general[0], setErrorMessage)
-            console.log('Error :', err?.response?.data?.message?.general[0])
-        })
-}
 export default ClubSetup
